+++ packages/antd/src/not-antd/rootOrDetails.tsx
@@ -1,8 +1,10 @@
+import { HookMap } from '@moxb/moxb';
+
 const uuidv4 = require('uuid/v4');
 import {
     getNextPathToken,
     isTokenEmpty,
-    Navigable,
+    NavigableContent,
     NavStateHooks,
     SubStateCoreInfo,
     TestLocation,
@@ -16,11 +18,12 @@ import { UIFragmentSpec } from './UIFragmentSpec';
  * This is the full spec of the root-or-details component
  */
 interface OwnProps<DataType> {
+    id: string;
     ifRoot: SubStateCoreInfo<UIFragmentSpec, DataType>;
     ifDetails: SubStateCoreInfo<UIFragmentSpec, DataType>;
 }
 
-type ComponentProps<DataType> = UsesLocation & Navigable<UIFragmentSpec, DataType>;
+type ComponentProps<DataType> = UsesLocation & NavigableContent<UIFragmentSpec, DataType>;
 
 export interface DetailProps {
     token: string;
@@ -30,11 +33,72 @@ export function rootOrDetails<DataType>(ownProps: OwnProps<DataType>) {
     const nodeId = uuidv4();
     return inject('locationManager')(
         observer((props: ComponentProps<DataType>) => {
-            let rootHooks: NavStateHooks | undefined;
-            const registerNavStateHooksForRoot = (hooks: NavStateHooks) => (rootHooks = hooks);
+            const { navControl } = props;
+            const rootHooks: HookMap = {};
+            const registerNavStateHooksForRoot = (componentId: string, hooks: NavStateHooks) =>
+                (rootHooks[componentId] = hooks);
+            const unregisterNavStateHooksForRoot = (componentId: string) => delete rootHooks[componentId];
+            const isRootActive = () => navControl.isActive() && isTokenEmpty(getNextPathToken(props));
+            const wouldRootBeActive = (testLocation: TestLocation) =>
+                navControl.wouldBeActive(testLocation) && isTokenEmpty(testLocation.pathTokens[props.parsedTokens!]);
 
-            let detailHooks: NavStateHooks | undefined;
-            const registerNavStateHooksForDetail = (hooks: NavStateHooks) => (detailHooks = hooks);
+            const detailHooks: HookMap = {};
+            const registerNavStateHooksForDetail = (componentId: string, hooks: NavStateHooks) =>
+                (detailHooks[componentId] = hooks);
+            const unregisterNavStateHooksForDetail = (componentId: string) => delete detailHooks[componentId];
+            const isDetailActive = () => navControl.isActive() && !isRootActive();
+            const wouldDetailBeActive = (testLocation: TestLocation) =>
+                navControl.wouldBeActive(testLocation) && !isTokenEmpty(testLocation.pathTokens[props.parsedTokens!]);
+
+            /**
+             * Compute which sub-states would be enabled when we changed from this older location
+             */
+            const getEnteredHooksFrom = (oldLocation: TestLocation): NavStateHooks[] => {
+                const oldRoot = wouldRootBeActive(oldLocation);
+                const oldDetail = wouldDetailBeActive(oldLocation);
+                const newRoot = isRootActive();
+                const newDetail = isDetailActive();
+
+                return [
+                    ...(!oldRoot && newRoot ? Object.values(rootHooks).filter(h => !!h) : []),
+                    ...(!oldDetail && newDetail ? Object.values(detailHooks).filter(h => !!h) : []),
+                ];
+            };
+
+            /**
+             * Compute which sub-states would be disabled if we changed to this new location
+             */
+            const getLeavingHooksTo = (newLocation: TestLocation): NavStateHooks[] => {
+                const oldActive = navControl.isActive();
+                const oldRoot = isRootActive();
+                const oldDetail = isDetailActive();
+                const newActive = navControl.wouldBeActive(newLocation);
+                const newRoot = wouldRootBeActive(newLocation);
+                const newDetail = wouldDetailBeActive(newLocation);
+                console.log(
+                    'Testing',
+                    newLocation,
+                    'for',
+                    ownProps.id,
+                    'active:',
+                    oldActive,
+                    '==>',
+                    newActive,
+                    'root:',
+                    oldRoot,
+                    '==>',
+                    newRoot,
+                    'detail:',
+                    oldDetail,
+                    '==>',
+                    newDetail,
+                    navControl
+                );
+                return [
+                    ...(oldRoot && !newRoot ? Object.values(rootHooks).filter(h => !!h) : []),
+                    ...(oldDetail && !newDetail ? Object.values(detailHooks).filter(h => !!h) : []),
+                ];
+            };
 
             /**
              * We register ourselves as a change interceptor,
@@ -52,38 +116,26 @@ export function rootOrDetails<DataType>(ownProps: OwnProps<DataType>) {
                  * This is our "change interceptor" hook, that will be called by the
                  * `LocationManager`.
                  */
-                // tslint:disable-next-line:cyclomatic-complexity
-                anyQuestionsFor(location: TestLocation): string[] {
-                    const oldToken = getNextPathToken(props);
-                    const oldRoot = isTokenEmpty(oldToken);
+                anyQuestionsFor(testLocation: TestLocation): string[] {
+                    return getLeavingHooksTo(testLocation)
+                        .map(h => h.getLeaveQuestion)
+                        .filter(gen => !!gen)
+                        .map(gen => gen!())
+                        .filter(q => !!q) as string[];
+                },
 
-                    const newToken = location.pathTokens[props.parsedTokens!];
-                    const newRoot = isTokenEmpty(newToken);
+                onBeforeChange(newLocation: TestLocation) {
+                    getLeavingHooksTo(newLocation)
+                        .map(h => h.onLeave)
+                        .filter(cb => !!cb)
+                        .forEach(cb => cb!());
+                },
 
-                    if (newRoot) {
-                        if (oldRoot) {
-                            // Staying at root, nothing to do.
-                        } else {
-                            // Going from detail to root.
-                            // We would hide the detail, so check with it.
-                            if (detailHooks && detailHooks.getLeaveQuestion) {
-                                const q = detailHooks.getLeaveQuestion();
-                                return q ? [q] : [];
-                            }
-                        }
-                    } else {
-                        if (oldRoot) {
-                            // Going from root to detail.
-                            // We would hide the root, so check with it.
-                            if (rootHooks && rootHooks.getLeaveQuestion) {
-                                const q = rootHooks.getLeaveQuestion();
-                                return q ? [q] : [];
-                            }
-                        } else {
-                            // Staying ad detail, nothing to do
-                        }
-                    }
-                    return [];
+                onAfterChange(oldLocation: TestLocation) {
+                    getEnteredHooksFrom(oldLocation)
+                        .map(h => h.onEnter)
+                        .filter(cb => !!cb)
+                        .forEach(cb => cb!());
                 },
             });
 
@@ -96,6 +148,9 @@ export function rootOrDetails<DataType>(ownProps: OwnProps<DataType>) {
                     checkCondition: true,
                     navControl: {
                         registerStateHooks: registerNavStateHooksForRoot,
+                        unregisterStateHooks: unregisterNavStateHooksForRoot,
+                        isActive: isRootActive,
+                        wouldBeActive: wouldRootBeActive,
                     },
                 });
             } else {
@@ -110,6 +165,9 @@ export function rootOrDetails<DataType>(ownProps: OwnProps<DataType>) {
                     checkCondition: true,
                     navControl: {
                         registerStateHooks: registerNavStateHooksForDetail,
+                        unregisterStateHooks: unregisterNavStateHooksForDetail,
+                        isActive: isDetailActive,
+                        wouldBeActive: wouldDetailBeActive,
                     },
                 });
             }