+++ packages/moxb/src/routing/location-state-space/state-space/StateSpaceHandlerImpl.ts
@@ -45,8 +45,12 @@ function filterSubStates<LabelType, WidgetType, DataType>(
         });
 }
 
-interface HookMap {
+export interface HookMap {
     [index: string]: NavStateHooks;
+}
+
+interface HookMapMap {
+    [index: string]: HookMap;
 }
 
 /**
@@ -63,7 +67,7 @@ export class StateSpaceHandlerImpl<LabelType, WidgetType, DataType>
     public readonly _subStatesInContext: SubStateInContext<LabelType, WidgetType, DataType>[];
     protected readonly _allSubStates: SubStateInContext<LabelType, WidgetType, DataType>[];
     protected readonly _filterCondition?: StateCondition<DataType>;
-    public readonly stateHooks: HookMap = {};
+    public readonly stateHooks: HookMapMap = {};
 
     /**
      * Add context info around a given sub-state
@@ -235,13 +239,31 @@ export class StateSpaceHandlerImpl<LabelType, WidgetType, DataType>
 
     registerNavStateHooksForSubState(
         subState: SubStateInContext<LabelType, WidgetType, DataType> | null,
+        componentId: string,
         hooks: NavStateHooks
     ) {
         if (!subState) {
             // console.log('Ignoring dirty tester for missing subState');
         } else {
-            // console.log('SubState', subState.menuKey, 'has state hooks', hooks);
-            this.stateHooks[subState.menuKey] = hooks;
+            const { menuKey: stateId } = subState;
+            if (!this.stateHooks[stateId]) {
+                this.stateHooks[stateId] = {};
+            }
+            this.stateHooks[stateId][componentId] = hooks;
+        }
+    }
+
+    unregisterNavStateHooksForSubState(
+        subState: SubStateInContext<LabelType, WidgetType, DataType> | null,
+        componentId: string
+    ) {
+        if (!subState) {
+            // console.log('Ignoring dirty tester for missing subState');
+        } else {
+            const { menuKey: stateId } = subState;
+            if (!!this.stateHooks[stateId] && !!this.stateHooks[stateId][componentId]) {
+                delete this.stateHooks[stateId][componentId];
+            }
         }
     }
 }